feat: refactor paste page into smaller components
